feat: move tests outside src directory
