+++ packages/after.js/test/components/AsyncNotFoundComponent.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 class AsyncNotFoundComponent extends React.Component {
   static displayName = 'AsyncNotFoundComponent';
-  static data = "The Page You Were Looking For Was Not Found"
+  static data = 'The Page You Were Looking For Was Not Found';
 
   static async getInitialProps() {
     return await Promise.resolve({ statusCode: 404 });
