+++ packages/after.js/test/components/AsyncRedirectComponent.tsx
