+++ packages/after.js/test/components/NoGetInitialProps.tsx
