+++ packages/after.js/test/components/NonAsyncRedirectComponent.tsx
