+++ packages/after.js/test/components/NonDynamicExport.tsx
